pyproject.toml
[tool.poetry]
name = "elapi"
version = "2.0.0"
description = "elAPI is a powerful, extensible API client for eLabFTW."
authors = [
    "Alexander Haller, Mahadi Xion <elabftw@uni-heidelberg.de>",
]
maintainers = [
    "Mahadi Xion <mahadi.xion@urz.uni-heidelberg.de>",
]
readme = "README.md"
homepage = "https://www.urz.uni-heidelberg.de/en/service-catalogue/software-and-applications/elabftw"
repository = "https://github.com/uhd-urz/elAPI/"
license = "AGPL-3.0-only"
keywords = ["elabftw", "api"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: AnyIO",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]

[project.urls]
Changelog = "https://github.com/uhd-urz/elAPI/blob/main/CHANGELOG.md"

[tool.poetry.scripts]
elapi = "elapi.cli.elapi:app"

[tool.poetry.dependencies]
python = "^3.9.0"
dynaconf = "^3.2.0"
httpx = "^0.24.1"
httpx-auth = "^0.17.0"
typer = { extras = ["all"], version = "^0.9.0" }
pyyaml = "^6.0.1"
mistune = "^3.0.2"
h2 = "^4.1.0"
python-dateutil = { version = "^2.8.2", optional = true }
tenacity = { version = "^8.2.3", optional = true }

[tool.poetry.extras]
uhd-urz = ["python-dateutil", "tenacity"]

[tool.poetry.group.dev.dependencies]
ptpython = "^3.0.23"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
